requirements.txt
pyspark==3.5.4
pandas==2.2.2
delta-spark==3.2.1
psycopg2-binary==2.9.9
SQLAlchemy==1.4.52
boto3==1.34.91
deltalake==0.22.3
polars==1.18.0
dag-factory==0.19.0
duckdb==1.2.0
python-dotenv==1.0.1